requirements.txt
langchain==0.2.15
langchain_community==0.2.13
langchain_groq==0.1.9
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.37.1
streamlit_extras==0.4.7
sentence_transformers
faiss-cpu